Get everything using the new camera

+++ SprintLevelEditor/SimpleRectangle.cs
@@ -23,7 +23,7 @@ namespace SprintLevelEditor
             Height = height;
         }
 
-        public static SimpleRectangle fromWall(Wall wall, int blockSize, int xOffset, int yOffset)
+        public static SimpleRectangle fromWall(Wall wall, float blockSize, int xOffset, int yOffset)
         {
             Rectangle rectangle = new Rectangle((int)wall.sprite.position.X, (int)wall.sprite.position.Y, (int)wall.sprite.DrawSize.Width, (int)wall.sprite.DrawSize.Height);
             return new SimpleRectangle(Math.Max(0, (rectangle.X - xOffset) / blockSize) * SCALE, Math.Max(0, (rectangle.Y - yOffset) / blockSize) * SCALE, Math.Max(1, rectangle.Width / blockSize) * SCALE, Math.Max(1, rectangle.Height / blockSize) * SCALE);
